2024/07/03

+++ recbole/evaluator/metrics.py
@@ -132,13 +132,13 @@ def ndcg_(pos_index, pos_len):
     len_rank = np.full_like(pos_len, pos_index.shape[1])
     idcg_len = np.where(pos_len > len_rank, len_rank, pos_len)
 
-    iranks = np.zeros_like(pos_index, dtype=np.float)
+    iranks = np.zeros_like(pos_index, dtype=np.cfloat)
     iranks[:, :] = np.arange(1, pos_index.shape[1] + 1)
     idcg = np.cumsum(1.0 / np.log2(iranks + 1), axis=1)
     for row, idx in enumerate(idcg_len):
         idcg[row, idx:] = idcg[row, idx - 1]
 
-    ranks = np.zeros_like(pos_index, dtype=np.float)
+    ranks = np.zeros_like(pos_index, dtype=np.cfloat)
     ranks[:, :] = np.arange(1, pos_index.shape[1] + 1)
     dcg = 1.0 / np.log2(ranks + 1)
     dcg = np.cumsum(np.where(pos_index, dcg, 0), axis=1)
